requirements.txt
aiogram==2.22.1
beautifulsoup4==4.11.1
bs4==0.0.1
ImageHash==4.3.0
opencv-python==4.6.0.66
Pillow==9.2.0
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.12
